Move `pants.backend.explorer` -> `pants.explorer.server` due to being in a dedicated distribution.
