+++ src/python/pants/explorer/server/graphql/context.py
@@ -8,8 +8,8 @@ from typing import cast
 
 from strawberry.types import Info
 
-from pants.backend.explorer.server.uvicorn import UvicornServer
 from pants.engine.explorer import RequestState
+from pants.explorer.server.uvicorn import UvicornServer
 
 
 @dataclass(frozen=True)
