+++ src/python/pants/explorer/server/graphql/query/rules_test.py
@@ -13,7 +13,7 @@ import pytest
             {"name": r"\.explorer\."},
             {
                 "rules": [
-                    {"name": "pants.backend.explorer.graphql.rules.get_graphql_uvicorn_setup"},
+                    {"name": "pants.explorer.server.graphql.rules.get_graphql_uvicorn_setup"},
                 ]
             },
         ),
@@ -21,7 +21,7 @@ import pytest
             {"name": r"\.graphql\."},
             {
                 "rules": [
-                    {"name": "pants.backend.explorer.graphql.rules.get_graphql_uvicorn_setup"},
+                    {"name": "pants.explorer.server.graphql.rules.get_graphql_uvicorn_setup"},
                 ]
             },
         ),
