+++ src/python/pants/explorer/server/graphql/query/targets.py
@@ -10,11 +10,11 @@ from typing import Iterable, Iterator, List, Optional
 import strawberry
 from strawberry.types import Info
 
-from pants.backend.explorer.graphql.context import GraphQLContext
-from pants.backend.explorer.graphql.field_types import JSONScalar
 from pants.backend.project_info.peek import TargetData, TargetDatas
 from pants.base.specs_parser import SpecsParser
 from pants.engine.target import AllUnexpandedTargets, UnexpandedTargets
+from pants.explorer.server.graphql.context import GraphQLContext
+from pants.explorer.server.graphql.field_types import JSONScalar
 from pants.help.help_info_extracter import TargetTypeHelpInfo
 from pants.util.strutil import softwrap
