+++ src/python/pants/explorer/server/graphql/rules.py
@@ -3,16 +3,16 @@
 
 from __future__ import annotations
 
-from pants.backend.explorer.browser import Browser, BrowserRequest
-from pants.backend.explorer.graphql.setup import graphql_uvicorn_setup
-from pants.backend.explorer.graphql.subsystem import GraphQLSubsystem
-from pants.backend.explorer.server.uvicorn import UvicornServerSetup, UvicornServerSetupRequest
 from pants.backend.project_info.peek import TargetDatas
 from pants.base.specs import Specs
 from pants.engine.environment import EnvironmentName
 from pants.engine.rules import Get, QueryRule, collect_rules, rule
 from pants.engine.target import AllUnexpandedTargets, UnexpandedTargets
 from pants.engine.unions import UnionRule
+from pants.explorer.server.browser import Browser, BrowserRequest
+from pants.explorer.server.graphql.setup import graphql_uvicorn_setup
+from pants.explorer.server.graphql.subsystem import GraphQLSubsystem
+from pants.explorer.server.uvicorn import UvicornServerSetup, UvicornServerSetupRequest
 
 
 class GraphQLUvicornServerSetupRequest(UvicornServerSetupRequest):
