+++ src/python/pants/explorer/server/uvicorn.py
@@ -10,12 +10,12 @@ from typing import Any, Callable
 from fastapi import FastAPI
 from uvicorn import Config, Server  # type: ignore
 
-from pants.backend.explorer.browser import BrowserRequest
 from pants.base.exiter import ExitCode
 from pants.engine.environment import EnvironmentName
 from pants.engine.explorer import ExplorerServer, ExplorerServerRequest, RequestState
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.unions import UnionMembership, union
+from pants.explorer.server.browser import BrowserRequest
 
 logger = logging.getLogger(__name__)
